requirements.txt
python-dotenv
azure-identity
fastapi
uvicorn
tiktoken
semantic-kernel[azure]
azure-ai-projects